move assumeReadType into its own file
