+++ mongodb/src/main/scala/quasar/physical/mongodb/planner/common.scala
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2014–2017 SlamData Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package quasar.physical.mongodb.planner
+
+import quasar.fs.{FileSystemError, MonadFsErr}
+
+object common {
+  def raiseErr[M[_], A](err: FileSystemError)(
+    implicit ev: MonadFsErr[M]
+  ): M[A] = ev.raiseError(err)
+
+  def handleErr[M[_], A](ma: M[A])(f: FileSystemError => M[A])(
+    implicit ev: MonadFsErr[M]
+  ): M[A] = ev.handleError(ma)(f)
+}
